[Update] 점수 관련 코드 제거 및 플레이어 상태 데이터 구조 개선
점수 업데이트 및 관련 코드와 UI 요소를 제거하고, 코드의 간결성을 높였습니다. 또한, 플레이어 상태 정보를 구조체로 분리하여 재사용성과 확장성을 강화했습니다.
