+++ Content/WeaponMaster/Blueprints/HUD/WBP/WBP_PlayerStatusWidgetExtended.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:84d9683ab7ac49dbe70d6bb5e9701763d4e56b740b2ead01ee50912952b13378
-size 41368
+oid sha256:1a649fb5661e512a80acae0c0b7b669d567e94251f63bf2b4396064b262bee1a
+size 41296
