+++ Source/WeaponMaster/UI/CommonUI/PlayerStatusWidget.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Blueprint/UserWidget.h"
+#include "Data/StatusTypes.h"
 #include "PlayerStatusWidget.generated.h"
 
 class UTextBlock;
@@ -9,42 +10,7 @@ class UProgressBar;
 class UImage;
 class UTexture2D;
 
-/** 플레이어 상태 정보 구조체 */
-USTRUCT(BlueprintType)
-struct FPlayerStatusInfo
-{
-    GENERATED_BODY()
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    FString PlayerName;
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    float CurrentHealth;
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    float MaxHealth;
-    
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    UTexture2D* PlayerThumbnailTexture;
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    int32 TeamID;
-    
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Status")
-    int32 CharacterID;
-    
-    // 생성자로 기본값 설정
-    FPlayerStatusInfo()
-        : PlayerName(TEXT("Unknown")),
-          CurrentHealth(100.0f),
-          MaxHealth(100.0f),
-          PlayerThumbnailTexture(nullptr),
-          TeamID(0),
-          CharacterID(-1)
-    {
-    }
-};
+
 
 UCLASS()
 class WEAPONMASTER_API UPlayerStatusWidget : public UUserWidget
