+++ Source/WeaponMaster/UI/MultiUI/PlayerStatusWidgetExtended.cpp
@@ -10,7 +10,6 @@ void UPlayerStatusWidgetExtended::NativeConstruct()
     // 초기값 설정
     Kills = 0;
     Deaths = 0;
-    Score = 0;
     bIsLocalPlayer = false;
     
     // 초기 텍스트 업데이트
@@ -29,24 +28,18 @@ void UPlayerStatusWidgetExtended::UpdateDeaths(int32 NewDeaths)
     UpdateTexts();
 }
 
-void UPlayerStatusWidgetExtended::UpdateScore(int32 NewScore)
-{
-    Score = NewScore;
-    UpdateTexts();
-}
-
 void UPlayerStatusWidgetExtended::SetIsLocalPlayer(bool bIsLocal)
 {
     bIsLocalPlayer = bIsLocal;
     
-    // 로컬 플레이어인 경우 시각적으로 강조
+    // If this is the local player, visually indicate it
     if (bIsLocalPlayer)
     {
-        // 부모 위젯에 Border 컴포넌트가 있다고 가정
+        // Find the border component if it exists in the widget hierarchy
         UBorder* BorderWidget = Cast<UBorder>(GetWidgetFromName(TEXT("PlayerBorder")));
         if (BorderWidget)
         {
-            // 로컬 플레이어를 강조하기 위한 색상 설정
+            // Highlight the local player with a distinct color
             BorderWidget->SetBrushColor(FLinearColor(0.1f, 0.5f, 1.0f, 0.3f));
         }
     }
@@ -69,15 +62,6 @@ void UPlayerStatusWidgetExtended::UpdateTexts()
         DeathsText->SetText(FText::Format(
             NSLOCTEXT("PlayerStatusExtended", "Deaths", "데스: {0}"),
             FText::AsNumber(Deaths)
-        ));
-    }
-    
-    // 점수 텍스트 업데이트
-    if (ScoreText)
-    {
-        ScoreText->SetText(FText::Format(
-            NSLOCTEXT("PlayerStatusExtended", "Score", "점수: {0}"),
-            FText::AsNumber(Score)
         ));
     }
 }
