+++ Source/WeaponMaster/UI/MultiUI/PlayerStatusWidgetExtended.h
@@ -7,8 +7,9 @@
 class UTextBlock;
 
 /**
- * 개인전 모드에서 사용할 확장된 플레이어 상태 위젯
- * 기본 PlayerStatusWidget을 상속받아 킬, 데스, 점수 등의 추가 정보를 표시합니다.
+ * Extended player status widget for use in different game modes
+ * Extends the base PlayerStatusWidget with additional features like
+ * kills, deaths tracking and local player highlighting.
  */
 UCLASS()
 class WEAPONMASTER_API UPlayerStatusWidgetExtended : public UPlayerStatusWidget
@@ -26,10 +27,6 @@ public:
     UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
     void UpdateDeaths(int32 NewDeaths);
     
-    /** 점수 업데이트 */
-    UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
-    void UpdateScore(int32 NewScore);
-    
     /** 킬 수 반환 */
     UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
     int32 GetKills() const { return Kills; }
@@ -38,24 +35,17 @@ public:
     UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
     int32 GetDeaths() const { return Deaths; }
     
-    /** 점수 반환 */
-    UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
-    int32 GetScore() const { return Score; }
-    
-    /** 내 플레이어인지 설정 */
+    /** Set whether this widget represents the local player */
     UFUNCTION(BlueprintCallable, Category = "PlayerStatus")
     void SetIsLocalPlayer(bool bIsLocal);
     
 protected:
-    // UI 구성요소 - BindWidget으로 블루프린트에서 연결
+    // UI components - to be bound with BindWidget in blueprint
     UPROPERTY(meta = (BindWidget))
     TObjectPtr<UTextBlock> KillsText;
     
     UPROPERTY(meta = (BindWidget))
     TObjectPtr<UTextBlock> DeathsText;
-    
-    UPROPERTY(meta = (BindWidget))
-    TObjectPtr<UTextBlock> ScoreText;
     
     // 데이터 속성
     UPROPERTY(BlueprintReadOnly, Category = "PlayerStatus")
@@ -64,9 +54,7 @@ protected:
     UPROPERTY(BlueprintReadOnly, Category = "PlayerStatus")
     int32 Deaths;
     
-    UPROPERTY(BlueprintReadOnly, Category = "PlayerStatus")
-    int32 Score;
-    
+    /** Whether this widget represents the local player */
     UPROPERTY(BlueprintReadOnly, Category = "PlayerStatus")
     bool bIsLocalPlayer;
